# develop/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Select Wave 選集 API 系統 - Swagger",
    "description": "部分API需求 \n注意事項：登入成功後請點「Authorize」輸入 Token。\n\n範例程式碼 :\n\n    fetch('/api/member/', { method: 'GET' })\n        .then(response => response.json())\n        .then(res => {\n            // { status: 'true', result: [{...}] }\n            console.log(res);\n        });\n    ",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "請加上 API Token"
    }
  },
  "paths": {
...
    "/api/imgur/": {
      "post": {
        "tags": [
          "Imgur - 圖片"
        ],
        "description": "上傳圖片",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "上傳圖片",
            "schema": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "object",
                  "properties": {
                    "image": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Success"
            },
            "description": "上傳成功"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "上傳失敗"
          },
          "401": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "未登入"
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "圖片格式錯誤"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "上傳失敗"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/member/": {
      "get": {
        "tags": [
          "Member - 會員"
        ],
        "description": "獲取所有會員",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "number",
            "description": "頁碼"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "description": "每頁記錄數"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Member"
            },
            "description": "獲取會員列表成功"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "獲取會員列表失敗"
          }
        }
      },
      "put": {
        "tags": [
          "Member - 會員"
        ],
        "description": "更新會員資料",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "更新會員資料",
            "schema": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "updateData": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Member"
            },
            "description": "更新會員資料成功"
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "找不到會員"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "更新會員資料失敗"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/poll/": {
      "get": {
        "tags": [
          "Poll - 投票"
        ],
        "description": "獲取所有投票",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "number",
            "description": "頁碼"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "description": "每頁記錄數"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Poll"
            },
            "description": "獲取投票列表成功"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "獲取投票列表失敗"
          }
        }
      },
      "post": {
        "tags": [
          "Poll - 投票"
        ],
        "description": "創建新投票",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "新增投票",
            "schema": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "description": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "imageUrl": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "tags": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            }
                          }
                        }
                      }
                    },
                    "userId": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "optionsData": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "properties": {
                              "type": "object",
                              "properties": {
                                "title": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    }
                                  }
                                },
                                "description": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    }
                                  }
                                },
                                "imageUrl": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Poll"
            },
            "description": "投票創建成功"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "投票創建失敗"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/comment/": {
      "post": {
        "tags": [
          "Comment - 評論"
        ],
        "description": "新增評論",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "新增評論",
            "schema": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Comment"
            },
            "description": "評論創建成功"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "評論創建失敗"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/option/": {
      "post": {
        "tags": [
          "Option - 投票及投票選項"
        ],
        "description": "新增選項",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "新增選項",
            "schema": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "description": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "imageUrl": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "pollId": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Option"
            },
            "description": "新增選項成功"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "新增選項失敗"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Option - 投票及投票選項"
        ],
        "description": "更新選項",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "更新選項",
            "schema": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "object",
                  "properties": {
                    "optionId": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "updateData": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Option"
            },
            "description": "更新選項成功"
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "找不到選項"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "更新選項失敗"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Poll": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62e348927278654321000004"
        },
        "title": {
          "type": "string",
          "example": "最喜歡的水果？"
        },
        "description": {
          "type": "string",
          "example": "請選擇您最喜歡的水果"
        },
        "imageUrl": {
          "type": "string",
          "example": "https://i.imgur.com/xcLTrkV.png"
        },
        "tags": {
          "type": "array",
          "example": [
            "水果",
            "投票"
          ],
          "items": {
            "type": "string"
          }
        },
        "createdBy": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "62e348927278654321000001"
            },
            "name": {
              "type": "string",
              "example": "John Doe"
            },
            "avatar": {
              "type": "string",
              "example": "https://i.imgur.com/xcLTrkV.png"
            }
          }
        },
        "createdTime": {
          "type": "string",
          "example": "2023-12-01T00:00:00.000Z"
        },
        "startDate": {
          "type": "string",
          "example": "2023-12-01T00:00:00.000Z"
        },
        "endDate": {
          "type": "string",
          "example": "2023-12-08T00:00:00.000Z"
        },
        "isPrivate": {
          "type": "boolean",
          "example": false
        },
        "totalVoters": {
          "type": "number",
          "example": 10
        },
        "like": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "62e348927278654321000005"
              },
              "pollId": {
                "type": "string",
                "example": "62e348927278654321000004"
              },
              "userId": {
                "type": "string",
                "example": "62e348927278654321000001"
              },
              "content": {
                "type": "string",
                "example": "我喜歡芒果！"
              },
              "createdTime": {
                "type": "string",
                "example": "2023-12-01T00:00:00.000Z"
              },
              "role": {
                "type": "string",
                "example": "user"
              },
              "edited": {
                "type": "boolean",
                "example": false
              },
              "updateTime": {
                "type": "string",
                "example": "2023-12-01T00:00:00.000Z"
              }
            }
          }
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "62e348927278654321000008"
              },
              "optionTitle": {
                "type": "string",
                "example": "芒果"
              },
              "optionImageUrl": {
                "type": "string",
                "example": "https://i.imgur.com/xcLTrkV.png"
              },
              "pollId": {
                "type": "string",
                "example": "62e348927278654321000004"
              }
            }
          }
        }
      }
    },
    "Option": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "637b24267525423242675259"
        },
        "optionTitle": {
          "type": "string",
          "example": "Red"
        },
        "optionImageUrl": {
          "type": "string",
          "example": "https://i.imgur.com/xcLTrkV.png"
        },
        "pollId": {
          "type": "string",
          "example": "637b24267525423242675257"
        },
        "voters": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "string",
                "example": "637b24267525423242675255"
              },
              "createdTime": {
                "type": "string",
                "example": "2023-11-16T07:52:54.232Z"
              },
              "updatedTime": {
                "type": "string",
                "example": "2023-11-16T07:52:54.232Z"
              }
            }
          }
        }
      }
    },
    "Comment": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "637b24267525423242675258"
        },
        "pollId": {
          "type": "string",
          "example": "637b24267525423242675257"
        },
        "userId": {
          "type": "string",
          "example": "637b24267525423242675255"
        },
        "content": {
          "type": "string",
          "example": "This is a great poll!"
        },
        "createdTime": {
          "type": "string",
          "example": "2023-11-16T07:52:54.232Z"
        },
        "role": {
          "type": "string",
          "example": "user"
        },
        "edited": {
          "type": "boolean",
          "example": false
        },
        "updateTime": {
          "type": "string",
          "example": "2023-11-16T07:52:54.232Z"
        }
      }
    }
  }
}